Moved FilmStore into the new Stores package.
